refactor: Restructure NLP service and remove legacy code
Clean up and reorganize the NLP service by:
- Removing the FreiFahren_BE_NLP directory and its contents
- Simplifying the main.py file
- Removing deprecated testing and NER-related files
- Preparing the service for future modularization

This change removes legacy code and reduces complexity in the NLP service while maintaining its core functionality.
