--- packages/nlp_service/FreiFahren_BE_NLP/.gitignore
@@ -1,9 +1,0 @@
-.env
-/__pycache__
-NER/__pycache__
-```
-.DS_Store
-
-.vscode/settings.json
-*.pyc
-*.log*
