--- packages/nlp_service/FreiFahren_BE_NLP/requirements.txt
@@ -1,18 +1,0 @@
-pyTelegramBotAPI
-fuzzywuzzy 
-python-Levenshtein
-python-dotenv
-numpy<2,>=1.26.0
-flake8
-flake8-quotes
-pep8-naming
-flake8-unused-arguments
-flake8-bugbear
-numpy<2,>=1.26.0
-
-watchdog
-openai
-
-spacy
-
-psycopg2-binary
