+++ packages/nlp_service/Pipfile
@@ -1,0 +1,28 @@
+[[source]]
+url = "https://pypi.org/simple"
+verify_ssl = true
+name = "pypi"
+
+[packages]
+waitress = "==3.0.2"
+pytelegrambotapi = "==4.26.0"
+fuzzywuzzy = "==0.18.0"
+python-levenshtein = "==0.26.1"
+python-dotenv = "==1.0.1"
+numpy = "==1.26.0"
+flake8 = "==7.1.1"
+flake8-quotes = "==3.4.0"
+pep8-naming = "==0.14.1"
+flake8-unused-arguments = "==0.0.13"
+flake8-bugbear = "==24.12.12"
+pytz = "==2025.1"
+watchdog = "==6.0.0"
+spacy = "==3.7.2"
+psycopg2-binary = "==2.9.9"
+flask = "==3.0.2"
+sentry-sdk = "==1.40.4"
+
+[dev-packages]
+
+[requires]
+python_version = "3.11"
