+++ packages/nlp_service/config/config.py
@@ -1,7 +1,6 @@
 import os
 from dotenv import load_dotenv
 
-
 load_dotenv()
 
 FREIFAHREN_CHAT_ID = os.getenv("FREIFAHREN_CHAT_ID")
@@ -14,3 +13,6 @@ REPORT_PASSWORD = os.getenv("REPORT_PASSWORD")
 # this is to restart the nlp service
 RESTART_PASSWORD = os.getenv("RESTART_PASSWORD")
 SENTRY_DSN = os.getenv("SENTRY_DSN")
+
+# minimum message length to be processed, for the core/processor.py
+MINIMUM_MESSAGE_LENGTH = 3
