+++ packages/nlp_service/core/data/synonyms.json
