+++ packages/nlp_service/core/testing/remove_direction_and_keyword_test.py
